requirements.txt
absl-py==1.0.0
altgraph==0.17.2
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
auto-py-to-exe==2.19.0
backcall==0.2.0
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
decorator==5.1.1
decord==0.6.0
distlib==0.3.4
docopt==0.6.2
Eel==0.12.4
executing==0.8.3
filelock==3.6.0
flatbuffers==2.0
future==0.18.2
gast==0.5.3
gevent==21.12.0
gevent-websocket==0.10.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h5py==3.6.0
idna==2.6
ipython==8.1.1
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.1.0
pipenv==2022.1.8
pipreqs==0.4.11
platformdirs==2.5.1
prompt-toolkit==3.0.28
protobuf==3.19.4
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-docs @ git+https://github.com/tensorflow/docs@65bd57a9f73429dfd798d17e0be0c73abbdb35fe
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tk==0.1.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.22
virtualenv==20.13.2
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.0.3
whichcraft==0.6.1
wrapt==1.13.3
yarg==0.1.9
zope.event==4.5.0
zope.interface==5.4.0
